+++ py_gql/utilities/ast_transforms.py
@@ -3,6 +3,7 @@
 Common utilities used to transform GraphQL documents.
 """
 
+from .._string_utils import camelcase_to_snakecase, snakecase_to_camelcase
 from ..lang.ast import Field
 from ..lang.visitor import DispatchingVisitor
 
@@ -15,4 +16,40 @@ class RemoveFieldAliasesVisitor(DispatchingVisitor):
     def enter_field(self, field: Field) -> Field:
         if field.alias is not None:
             field.alias = None
+        return field
+
+
+class CamelCaseToSnakeCaseVisitor(DispatchingVisitor):
+    """
+    Visitor implementation which renames field from camelCase to snake_case.
+
+    This is useful when working between languages with different conventions
+    such as Python and Javascript.
+
+    Note:
+        This only work on the incoming document and usually needs to be paired
+        with a post-processing step on the client or before sending out the
+        response.
+    """
+
+    def enter_field(self, field: Field) -> Field:
+        field.name.value = camelcase_to_snakecase(field.name.value)
+        return field
+
+
+class SnakeCaseToCamelCaseVisitor(DispatchingVisitor):
+    """
+    Visitor implementation which renames field from snake_case to camelCase.
+
+    This is useful when working between languages with different conventions
+    such as Python and Javascript.
+
+    Note:
+        This only work on the incoming document and usually needs to be paired
+        with a post-processing step on the client or before sending out the
+        response.
+    """
+
+    def enter_field(self, field: Field) -> Field:
+        field.name.value = snakecase_to_camelcase(field.name.value)
         return field