+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - (Re) add support for `copy` and `deepcopy` to `py_gql.lang.ast.Node`.
 - Add `py_gql.utilities.ast_transforms.RemoveFieldAliasesVisitor`.
+- Add `py_gql.utilities.ast_transforms.CamelCaseToSnakeCaseVisitor` and `py_gql.utilities.ast_transforms.SnakeCaseToCamelCaseVisitor`.
 
 ### Fixed